requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
torch>=1.10.0
transformers>=4.15.0
sentence-transformers>=2.2.0

# Vector database
faiss-cpu>=1.7.0
# faiss-gpu>=1.7.0  # Uncomment for GPU support
chromadb>=0.4.0

# LLM integration
openai>=1.0.0
langchain>=0.0.200

# Text processing
nltk>=3.6.0
spacy>=3.2.0
beautifulsoup4>=4.10.0
python-docx>=0.8.11
PyPDF2>=2.0.0

# API and web
fastapi>=0.95.0
uvicorn>=0.15.0
pydantic>=1.9.0

# Utilities
tqdm>=4.62.0
pyyaml>=6.0
python-dotenv>=0.19.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Documentation
sphinx>=4.3.0 